Send telemetry record on MCE
If CLEAR_TELEM is defined at compile time, and libtelemetry exists
at run time, telemetry records are enabled. When an MCE is detected,
a telemetry record is sent to the telemetry server defined in the
user's telemetrics configuration.

mcelog continues to log in it's normal manner (syslog, logfile).

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>

fix review